add child component and formatting
